requirements.txt

dagster
dagster-webserver
yfinance
psycopg2-binary
python-dotenv
